+++ server/controllers/test-suite.controller.ts
@@ -1,13 +1,13 @@
-import EndpointService from "../../services/endpoint/endpoint.service";
-import { TestSuiteService } from "../../services/testSuite/testSuite.service";
-import { TestService } from "../../services/testSuite/test.service";
+import EndpointService from "../services/endpoint/endpoint.service";
+import { TestSuiteService } from "../services/testSuite/testSuite.service";
+import { TestService } from "../services/testSuite/test.service";
 import { NextFunction, Request, Response, Router } from "express";
-import { Rule } from "../../model/testSuite/types";
-import { RequestService } from "../../services/request/request.service";
-import { SnifferService } from "../../services/sniffer/sniffer.service";
-import { useLog } from "../../lib/log";
-import { TestExecutionService } from "../../services/testSuite/testExecution.service";
-import { TestExecutor } from "./test-executor";
+import { Rule } from "../model/testSuite/types";
+import { RequestService } from "../services/request/request.service";
+import { SnifferService } from "../services/sniffer/sniffer.service";
+import { useLog } from "../lib/log";
+import { TestExecutionService } from "../services/testSuite/testExecution.service";
+import { TestExecutor } from "../services/testSuite/test-executor.service";
 
 const log = useLog({
   dirname: __dirname,
@@ -22,7 +22,7 @@ export class TestSuiteController {
     private readonly testService: TestService,
     private readonly requestService: RequestService,
     private readonly snifferService: SnifferService,
-    private readonly testExecutionService: TestExecutionService,
+    private readonly testExecutionService: TestExecutionService
   ) {
     this.testExecutor = new TestExecutor(this.requestService);
   }
@@ -43,7 +43,7 @@ export class TestSuiteController {
         const userId = res.locals.auth.user.id;
         const testSuites = await this.testSuiteService.getByUserId(userId);
         res.json(testSuites);
-      }),
+      })
     );
 
     router.post(
@@ -53,7 +53,7 @@ export class TestSuiteController {
         const { name } = req.body;
         const testSuite = await this.testSuiteService.create(name, userId);
         res.status(201).json(testSuite);
-      }),
+      })
     );
 
     router.post(
@@ -65,7 +65,7 @@ export class TestSuiteController {
           const { invocationId, testSuiteId } = req.params;
           const invocation = await this.endpointService.getInvocationById(
             invocationId,
-            userId,
+            userId
           );
 
           if (!invocation) {
@@ -74,7 +74,7 @@ export class TestSuiteController {
 
           const headerRules: Rule[] = Object.entries(
             // @ts-ignore
-            invocation?.response?.headers || {},
+            invocation?.response?.headers || {}
           ).map(([key, value]) => ({
             type: "header",
             comparator: "equals",
@@ -104,13 +104,13 @@ export class TestSuiteController {
                 targetPath: "",
               },
               ...headerRules,
-            ],
+            ]
           );
           res.status(201).json(test);
         } catch (error) {
           res.status(500).send();
         }
-      }),
+      })
     );
 
     router.delete(
@@ -118,8 +118,9 @@ export class TestSuiteController {
       catchAsync(async (req: Request, res: Response) => {
         try {
           const { testSuiteId, testId } = req.params;
-          const testSuite =
-            await this.testService.getByTestSuiteId(testSuiteId);
+          const testSuite = await this.testService.getByTestSuiteId(
+            testSuiteId
+          );
           if (!testSuite) {
             return res.status(404).send();
           }
@@ -136,7 +137,7 @@ export class TestSuiteController {
           log.error(e);
           res.status(500).send();
         }
-      }),
+      })
     );
 
     router.get(
@@ -152,7 +153,7 @@ export class TestSuiteController {
           return res.status(404).send();
         }
         res.json(test);
-      }),
+      })
     );
 
     router.get(
@@ -164,7 +165,7 @@ export class TestSuiteController {
           return res.status(404).send();
         }
         res.json(testSuite);
-      }),
+      })
     );
 
     router.put(
@@ -178,7 +179,7 @@ export class TestSuiteController {
         const { name } = req.body;
         await this.testSuiteService.update(testSuiteId, name);
         res.status(204).send();
-      }),
+      })
     );
 
     router.delete(
@@ -191,7 +192,7 @@ export class TestSuiteController {
         }
         await this.testSuiteService.deleteById(testSuiteId);
         res.status(204).send();
-      }),
+      })
     );
 
     router.put(
@@ -217,7 +218,7 @@ export class TestSuiteController {
           method,
         });
         res.status(204).send();
-      }),
+      })
     );
 
     router.post(
@@ -227,8 +228,9 @@ export class TestSuiteController {
           const { testSuiteId, testId } = req.params;
           const userId = res.locals.auth.user.id;
 
-          const testSuite =
-            await this.testService.getByTestSuiteId(testSuiteId);
+          const testSuite = await this.testService.getByTestSuiteId(
+            testSuiteId
+          );
           if (!testSuite) {
             return res.status(404).send();
           }
@@ -240,7 +242,7 @@ export class TestSuiteController {
 
           const sniffer = await this.snifferService.getSniffer(
             userId,
-            test.snifferId,
+            test.snifferId
           );
           if (!sniffer) {
             return res.status(404).send();
@@ -251,7 +253,7 @@ export class TestSuiteController {
           const results = await this.testExecutor.execute(
             test,
             sniffer.subdomain,
-            testExecution.id,
+            testExecution.id
           );
 
           await this.testExecutionService.update(testExecution.id, results);
@@ -261,7 +263,7 @@ export class TestSuiteController {
           log.error(e);
           res.status(500).send();
         }
-      }),
+      })
     );
 
     router.get(
@@ -273,7 +275,7 @@ export class TestSuiteController {
           (await this.testService.getByUrl(testSuiteId, url as string)) || [];
 
         const testExecutions = await this.testExecutionService.getByTestId(
-          tests.map((test) => test.id),
+          tests.map((test) => test.id)
         );
 
         if (!testExecutions) {
@@ -299,7 +301,7 @@ export class TestSuiteController {
         }
 
         res.json(results);
-      }),
+      })
     );
 
     router.get(
@@ -345,7 +347,7 @@ export class TestSuiteController {
         }
 
         res.json(results);
-      }),
+      })
     );
 
     return { path: "/sharkio/test-suites", router };
