+++ server/main.ts
@@ -8,7 +8,7 @@ import { InvocationController } from "./controllers/invocation.controller";
 import SettingsController from "./controllers/settings";
 import { SnifferController } from "./controllers/sniffer.controller";
 import { MockController } from "./controllers/mock.controller";
-import { TestSuiteController } from "./controllers/tests/test-suite.controller";
+import { TestSuiteController } from "./controllers/test-suite.controller";
 import { SwaggerUiController } from "./lib/swagger/swagger-controller";
 import ApiKeyRepository from "./model/apikeys/apiKeys.model";
 import ChatRepository from "./model/chat/chat.model";
@@ -80,7 +80,7 @@ async function main() {
   const responseService = new ResponseService(responseRepository);
   const endpointService = new EndpointService(
     endpointRepository,
-    invocationRepository,
+    invocationRepository
   );
   const userService = new UserService(userRepository);
   const apiKeyService = new APIKeysService(apiKeyRepository, userRepository);
@@ -90,7 +90,7 @@ async function main() {
   const testService = new TestService(testRepository);
   const requestService = new RequestService(invocationRepository);
   const testExecutionService = new TestExecutionService(
-    testExecutionRepository,
+    testExecutionRepository
   );
   const importService = new ImportService(endpointService);
 
@@ -101,25 +101,25 @@ async function main() {
   const cliController = new CLIController(
     apiKeyService,
     userService,
-    snifferService,
+    snifferService
   );
   const snifferController = new SnifferController(
     snifferService,
     docGenerator,
     endpointService,
-    mockService,
+    mockService
   );
   const endpointController = new EndpointController(
     endpointService,
     snifferService,
     requestService,
-    importService,
+    importService
   );
   const invocationController = new InvocationController(endpointService);
   const chatController = new ChatController(
     snifferService,
     endpointService,
-    chatService,
+    chatService
   );
   const swaggerUi = new SwaggerUiController();
   const testSuiteController = new TestSuiteController(
@@ -128,30 +128,30 @@ async function main() {
     testService,
     requestService,
     snifferService,
-    testExecutionService,
+    testExecutionService
   );
 
   /* Middlewares */
   const requestInterceptorMiddleware = new RequestInterceptor(
     snifferService,
     endpointService,
-    responseService,
+    responseService
   );
   const proxyMiddleware = new ProxyMiddleware(
     snifferService,
-    requestInterceptorMiddleware,
+    requestInterceptorMiddleware
   );
   const mockMiddleware = new MockMiddleware(
     mockService,
     snifferService,
-    responseService,
+    responseService
   );
 
   /* Servers */
   const proxyServer = new ProxyServer(
     proxyMiddleware,
     requestInterceptorMiddleware,
-    mockMiddleware,
+    mockMiddleware
   );
   const snifferManagerServer = new Server(
     [
@@ -165,7 +165,7 @@ async function main() {
       testSuiteController.getRouter(),
       mockController.getRouter(),
     ],
-    swaggerUi,
+    swaggerUi
   );
 
   // /* Start Servers */
