+++ server/model/testSuite/types.ts
@@ -1,5 +1,13 @@
 type RuleType = "status_code" | "body" | "header";
-type RuleComparator = "equals";
+type RuleComparator =
+  | "equals"
+  | "not_equals"
+  | "contains"
+  | "not_contains"
+  | "gt"
+  | "lt"
+  | "gte"
+  | "lte";
 export type Rule = {
   type: RuleType;
   comparator: RuleComparator;
