+++ server/package.json
@@ -63,6 +63,7 @@
     "@types/express-http-proxy": "^1.6.3",
     "@types/jest": "^29.5.7",
     "@types/jsonwebtoken": "^9.0.5",
+    "@types/lodash": "^4.14.202",
     "@types/node": "^20.8.7",
     "@types/random-string": "^0.0.31",
     "@types/supertest": "^2.0.12",
